Run prepull only if PREPULL_YAML is specified

+++ gce/run-e2e.sh
@@ -28,23 +28,26 @@ for node in $WINDOWS_NODES; do
   kubectl taint node $node node.kubernetes.io/os:NoSchedule-
 done
 
-# Pre-pull all the test images. The images are currently hard-coded.
-# Eventually, we should get the list directly from
-# https://github.com/kubernetes/kubernetes/blob/master/test/utils/image/manifest.go.
-PREPULL_FILE=${PREPULL_YAML:-prepull-head.yaml}
-SCRIPT_ROOT=$(cd `dirname $0` && pwd)
-kubectl create -f ${SCRIPT_ROOT}/${PREPULL_FILE}
-# Wait a while for the test images to be pulled onto the nodes. In empirical
-# testing it could take up to 30 minutes to finish pulling all the test
-# containers on a node.
-timeout ${PREPULL_TIMEOUT:-30m} kubectl wait --for=condition=ready pod -l prepull-test-images=e2e --timeout -1s
-# Check the status of the pods.
-kubectl get pods -o wide
-kubectl describe pods
-# Delete the pods anyway since pre-pulling is best-effort
-kubectl delete -f ${SCRIPT_ROOT}/${PREPULL_FILE}
-# Wait a few more minutes for the pod to be cleaned up.
-timeout 3m kubectl wait --for=delete pod -l prepull-test-images=e2e --timeout -1s
+# Run prepull if PREPULL_YAML is set to a value.
+if [[ -v PREPULL_YAML && ! -z "$PREPULL_YAML" ]]; then
+  # Pre-pull all the test images. The images are currently hard-coded.
+  # Eventually, we should get the list directly from
+  # https://github.com/kubernetes/kubernetes/blob/master/test/utils/image/manifest.go.
+  PREPULL_FILE=${PREPULL_YAML:-prepull-head.yaml}
+  SCRIPT_ROOT=$(cd `dirname $0` && pwd)
+  kubectl create -f ${SCRIPT_ROOT}/${PREPULL_FILE}
+  # Wait a while for the test images to be pulled onto the nodes. In empirical
+  # testing it could take up to 30 minutes to finish pulling all the test
+  # containers on a node.
+  timeout ${PREPULL_TIMEOUT:-30m} kubectl wait --for=condition=ready pod -l prepull-test-images=e2e --timeout -1s
+  # Check the status of the pods.
+  kubectl get pods -o wide
+  kubectl describe pods
+  # Delete the pods anyway since pre-pulling is best-effort
+  kubectl delete -f ${SCRIPT_ROOT}/${PREPULL_FILE}
+  # Wait a few more minutes for the pod to be cleaned up.
+  timeout 3m kubectl wait --for=delete pod -l prepull-test-images=e2e --timeout -1s
+fi
 
 # Download and set the list of test image repositories to use.
 curl \
